requirements.txt
Flask==1.0.3
Bootstrap-Flask==1.4
https://download.pytorch.org/whl/cpu/torch-1.0.0-cp37-cp37m-linux_x86_64.whl
torchvision==0.2.1
numpy==1.16.4
Pillow==6.2.2
gunicorn==19.9.0
requests==2.23.0